[Build] Finish the tree-sitter 0.27 index and capture migration
Child indexes are already u32 in 0.27, so the seven remaining casts are
now redundant and denied by lint. Query match captures moved from a
public field to an accessor returning a slice.

Verified with the repository lint and test targets on the pinned
toolchain, which is what caught the accessor returning a slice rather
than an iterator.

+++ crates/harn-hostlib/src/ast/edit_common.rs
@@ -263,7 +263,7 @@ pub(super) fn first_syntax_error(source: &[u8], language: Language) -> Option<St
             ));
         }
         for i in (0..node.child_count()).rev() {
-            if let Some(child) = node.child(i as u32) {
+            if let Some(child) = node.child(i) {
                 if child.has_error() || child.is_missing() {
                     stack.push(child);
                 }
